+++ fhir_path/lib/utils/utils.dart
@@ -1,4 +1,5 @@
 export 'deep_comparison_lists.dart';
+export 'exceptions.dart';
 export 'fhir_datatypes.dart';
 export 'polymorphic_prefixes.dart';
 export 'versions.dart';